refactor: move options to server.go, remove defaults method
