+++ server/server.go
@@ -12,13 +12,26 @@ type server struct {
 	log logger
 }
 
+// Options holds the configuration for the server.
+type Options struct {
+	Log logger
+}
+
+// Option is a function that configures the server.
+type Option func(*server)
+
 // New returns a new server.
 func New(options ...Option) *server {
 	s := &server{}
 	for _, option := range options {
 		option(s)
 	}
-	return s.defaults()
+
+	if s.log == nil {
+		s.log = NewDefaultLogger()
+	}
+
+	return s
 }
 
 // Start the server.
@@ -57,11 +70,10 @@ func (s server) shutdown() (os.Signal, error) {
 	return sig, nil
 }
 
-// defaults sets the default values for the server if they are not set.
-func (s *server) defaults() *server {
-	// Server default setup here.
-	if s.log == nil {
-		s.log = NewDefaultLogger()
+// WithOptions configures the server with the given Options.
+func WithOptions(options Options) Option {
+	return func(s *server) {
+		// Setup on all options from the option struct here.
+		s.log = options.Log
 	}
-	return s
 }
